# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baskonia Pipeline API",
    "description": "API de análisis de Baskonia (fase F3 de la migración strangler-fig).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/teams/{slug}/roster": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Roster",
        "description": "Plantilla actual de un equipo con la forma reciente de cada jugador (endpoint 8).",
        "operationId": "get_roster_api_v1_teams__slug__roster_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Año de inicio de temporada (p.ej. 2025)",
              "title": "Season"
            },
            "description": "Año de inicio de temporada (p.ej. 2025)"
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Competición: acb | euroleague | supercopa",
              "title": "League"
            },
            "description": "Competición: acb | euroleague | supercopa"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{slug}/players/form": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Player Form",
        "description": "Forma reciente por jugador (endpoint 9).",
        "operationId": "get_player_form_api_v1_teams__slug__players_form_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          },
          {
            "name": "last_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Last N"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Año de inicio de temporada (p.ej. 2025)",
              "title": "Season"
            },
            "description": "Año de inicio de temporada (p.ej. 2025)"
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Competición: acb | euroleague | supercopa",
              "title": "League"
            },
            "description": "Competición: acb | euroleague | supercopa"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerFormResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{slug}/players/streaks": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Streaks",
        "description": "Rachas de los jugadores en una temporada (endpoint 10).",
        "operationId": "get_streaks_api_v1_teams__slug__players_streaks_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          },
          {
            "name": "recent_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Recent N"
            }
          },
          {
            "name": "min_season_games",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 6,
              "title": "Min Season Games"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Año de inicio de temporada (p.ej. 2025)",
              "title": "Season"
            },
            "description": "Año de inicio de temporada (p.ej. 2025)"
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Competición: acb | euroleague | supercopa",
              "title": "League"
            },
            "description": "Competición: acb | euroleague | supercopa"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreaksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{slug}/players/load": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Load",
        "description": "Carga de minutos por jugador en la ventana de días (endpoint 11).",
        "operationId": "get_load_api_v1_teams__slug__players_load_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          },
          {
            "name": "window_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 14,
              "title": "Window Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/teams/{slug}/reports/scouting.pdf": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Team Report",
        "description": "Informe de equipo (endpoint 17) — implementado en F6.",
        "operationId": "team_report_api_v1_teams__slug__reports_scouting_pdf_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportNotImplemented"
                }
              }
            }
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportNotImplemented"
                }
              }
            },
            "description": "Not Implemented"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{slug}/reports/roster.pptx": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Matchup Report",
        "description": "Informe de enfrentamiento (endpoint 18) — implementado en F6.",
        "operationId": "matchup_report_api_v1_teams__slug__reports_roster_pptx_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportNotImplemented"
                }
              }
            }
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportNotImplemented"
                }
              }
            },
            "description": "Not Implemented"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoadItem": {
        "properties": {
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "games": {
            "type": "integer",
            "title": "Games"
          },
          "total_minutes": {
            "type": "number",
            "title": "Total Minutes"
          },
          "avg_minutes": {
            "type": "number",
            "title": "Avg Minutes"
          }
        },
        "type": "object",
        "required": [
          "player_name",
          "games",
          "total_minutes",
          "avg_minutes"
        ],
        "title": "LoadItem",
        "description": "Carga de minutos de un jugador en la ventana."
      },
      "LoadResponse": {
        "properties": {
          "window_days": {
            "type": "integer",
            "title": "Window Days"
          },
          "games_in_window": {
            "type": "integer",
            "title": "Games In Window"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LoadItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "window_days",
          "games_in_window",
          "note",
          "items"
        ],
        "title": "LoadResponse",
        "description": "Carga de minutos por jugador (transversal a temporada/competición)."
      },
...
      "PlayerFormItem": {
        "properties": {
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "games": {
            "type": "integer",
            "title": "Games"
          },
          "avg_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Minutes"
          },
          "avg_pts": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Pts"
          },
          "avg_pts_per36": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Pts Per36"
          },
          "avg_efg_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Efg Pct"
          },
          "avg_ts_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Ts Pct"
          },
          "avg_plus_minus": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Plus Minus"
          },
          "avg_turnovers": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Turnovers"
          },
          "fg3a_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fg3A Rate"
          },
          "ft_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ft Rate"
          }
        },
        "type": "object",
        "required": [
          "player_name",
          "games"
        ],
        "title": "PlayerFormItem",
        "description": "Fila de forma reciente por jugador (mapeo 1:1 con player_recent_form)."
      },
      "PlayerFormResponse": {
        "properties": {
          "last_n": {
            "type": "integer",
            "title": "Last N"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlayerFormItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "last_n",
          "items"
        ],
        "title": "PlayerFormResponse",
        "description": "Forma reciente por jugador."
      },
...
      "ReportNotImplemented": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ReportNotImplemented",
        "description": "Respuesta de los endpoints de informes en F3 (se completan en F6)."
      },
      "RosterPlayer": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "form": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RosterPlayer",
        "description": "Jugador de la plantilla actual con su forma reciente."
      },
      "RosterResponse": {
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamRef"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/RosterPlayer"
            },
            "type": "array",
            "title": "Players"
          }
        },
        "type": "object",
        "required": [
          "team",
          "players"
        ],
        "title": "RosterResponse",
        "description": "Plantilla actual de un equipo."
      },
...
      "StreakItem": {
        "properties": {
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "games_season": {
            "type": "integer",
            "title": "Games Season"
          },
          "recent_avg_pts": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Avg Pts"
          },
          "season_avg_pts": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Avg Pts"
          },
          "season_std_pts": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Std Pts"
          },
          "z_score_pts": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Z Score Pts"
          },
          "recent_avg_ts_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Avg Ts Pct"
          },
          "season_avg_ts_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Avg Ts Pct"
          },
          "season_std_ts_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Std Ts Pct"
          },
          "z_score_ts": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Z Score Ts"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "player_name",
          "games_season",
          "label"
        ],
        "title": "StreakItem",
        "description": "Racha de un jugador dentro de una temporada."
      },
      "StreaksResponse": {
        "properties": {
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "recent_n": {
            "type": "integer",
            "title": "Recent N"
          },
          "min_season_games": {
            "type": "integer",
            "title": "Min Season Games"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StreakItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "season",
          "recent_n",
          "min_season_games",
          "items"
        ],
        "title": "StreaksResponse",
        "description": "Rachas de los jugadores de un equipo en una temporada."
      },
...
      "TeamRef": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name"
        ],
        "title": "TeamRef",
        "description": "Referencia a un equipo (slug + nombre para mostrar)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}